+++ frontend/srcs/src/app/components/spectate/result-card-player/result-card-player.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+import { User } from 'src/app/models';
+
+@Component({
+  selector: 'app-result-card-player',
+  templateUrl: './result-card-player.component.html',
+  styleUrls: ['./result-card-player.component.scss']
+})
+export class ResultCardPlayerComponent {
+  @Input() user!: User;
+
+}